voting-app/vote/requirements.txt
Flask
Redis
gunicorn
grpcio
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-requests